Fix-up to Handle Breaking BigDecimal Change in Swagger 1.5.11
Methods in `AbstractNumericProperty` that used to take-in doubles now take in `BigDecimal` objects instead (way to follow semver guys...).

More here: https://github.com/swagger-api/swagger-core/issues/1889

+++ rest/api-descriptor/src/test/java/org/forgerock/api/transform/OpenApiTransformerTest.java
@@ -12,6 +12,7 @@
  * information: "Portions copyright [year] [name of copyright owner]".
  *
  * Copyright 2016 ForgeRock AS.
+ * Portions Copyright 2018 Wren Security.
  */
 package org.forgerock.api.transform;
 
@@ -31,6 +32,7 @@ import static org.forgerock.json.JsonValue.json;
 import static org.forgerock.json.JsonValue.object;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -561,8 +563,8 @@ public class OpenApiTransformerTest {
                             @Override
                             public Property get() {
                                 final LocalizableLongProperty o = new LocalizableLongProperty();
-                                o.setMinimum(1.0);
-                                o.setMaximum(2.0);
+                                o.setMinimum(BigDecimal.valueOf(1.0d));
+                                o.setMaximum(BigDecimal.valueOf(2.0d));
                                 o.setExclusiveMinimum(true);
                                 o.setExclusiveMaximum(true);
                                 o.setReadOnly(true);
@@ -591,8 +593,8 @@ public class OpenApiTransformerTest {
                             @Override
                             public Property get() {
                                 final LocalizableDoubleProperty o = new LocalizableDoubleProperty();
-                                o.setMinimum(1.0);
-                                o.setMaximum(2.0);
+                                o.setMinimum(BigDecimal.valueOf(1.0d));
+                                o.setMaximum(BigDecimal.valueOf(2.0d));
                                 o.setExclusiveMinimum(true);
                                 o.setExclusiveMaximum(true);
                                 o.setDefault(1.1);
